split predict.go into buckets.go and plan.go
Separation of concerns: predict.go was mixing two unrelated
responsibilities — bucket modeling/routing (data plumbing, touched
when adding routing rules) and the scaling math itself (the
algorithm, touched when tuning scaling logic). They evolve at
different rates and reading one to understand the other was noise.

  buckets.go  agentBucket / bucketState / bucketDecision types,
              agentBuckets, routeTaskToBucket, matchAgentToBucket,
              computeBucketStates
  plan.go     rawDelta, planScaling, allocateBudget

Also extracts the budget-allocation block of planScaling into its
own allocateBudget function. Reads top-to-bottom now: planScaling
reads as 'compute states, compute raw deltas, allocate budget,
emit decisions'. The math hidden in the middle of the loop is gone.
